Cleanup basics rust
