--- rust/basics/src/main.rs
@@ -1,4 +1,0 @@
-mod p0_durable_execution;
-mod p1_building_blocks;
-mod p2_virtual_objects;
-mod p3_workflows;
