--- rust/basics/src/utils/mod.rs
@@ -1,5 +1,0 @@
-mod stubs;
-pub mod subscription_service;
-
-pub use stubs::*;
-pub use subscription_service::SubscriptionServiceClient;
